split actual latex generation to _diagram.py
